Fix hunyuan_video_1_5_i2v_720p_gguf: use float timestep to avoid Long/Float dtype mismatch

+++ hunyuan_video_1_5_i2v_720p_gguf/pytorch/loader.py
@@ -226,7 +226,9 @@ class ModelLoader(ForgeModel):
             config.image_embed_dim,
             dtype=dtype,
         )
-        timestep = torch.tensor([500], dtype=torch.long).expand(batch_size)
+        # HunyuanVideo15TimestepEmbedder casts sinusoidal projections back to
+        # timestep.dtype; passing Long would produce Long→Float dtype mismatch.
+        timestep = torch.tensor([500], dtype=dtype).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
